Adding headers, cleaning up some small things
