+++ examples/gnn_cost_model/train_cost_model_test.py
@@ -1,9 +1,12 @@
-"""Tests for //compiler_gym/bin:train.py"""
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""Tests for examples/gnn_cost_model/train_cost_model_test.py"""
 
 
 import sys
-
-sys.path.append("./examples/gnn_cost_model")
 
 from absl import flags
 from train_cost_model import main
@@ -17,6 +20,7 @@ def test_run_train_smoke_test():
     flags = [
         "argv0",
         "--dataset_size=64",
+        "--batch_size=4",
         "--num_epoch=2",
     ]
     sys.argv = flags
