--- examples/gnn_cost_model/utils/parsing_utils.py
@@ -1,94 +1,0 @@
-import pickle
-import torch
-import dgl
-
-
-def load_vocab(load_pth):
-    return pickle.load(open(load_pth, "rb"))
-
-
-def process_networkx_graph(
-    graph,
-    vocab,
-    node_feature_list=["text", "type"],
-    edge_feature_list=["flow", "position"],
-):
-    """
-    Handles all of the requirements of taking a networkx graph and converting it into a
-    dgl graph
-    Inputs:
-        - graph: the networkx graph
-        - vocab: the vocabulary, a mapping from word to index.
-        - node_feature_list: a list of textual features from the networkx node that we want to make sure
-            are featurizable into a vector.
-        - edge_feature_list: a list of textual features from the networkx edges that we want to make sure
-            are featurizable into a vector.
-    """
-    update_graph_with_vocab(graph.nodes, node_feature_list, vocab)
-    update_graph_with_vocab(graph.edges, edge_feature_list, vocab)
-
-    dgl_graph = fast_networkx_to_dgl(graph)
-    return dgl_graph
-
-
-def fast_networkx_to_dgl(
-    graph, node_attrs=["text_idx", "type"], edge_attrs=["flow", "position"]
-):
-    """
-    Takes a networkx graph and its given node attributes and edge attributes
-    and converts it corresponding dgl graph
-    Inputs:
-        - graph: the networkx graph
-        - node_attrs: node attributes to convert
-        - edge_attrs: edge attributes to convert
-    """
-
-    edges = [edge for edge in graph.edges()]
-    dgl_graph = dgl.graph(edges, num_nodes=graph.number_of_nodes())
-
-    for feat in edge_attrs:
-        edge_assigns = torch.tensor(
-            [val[-1] for val in graph.edges(data=feat)], dtype=torch.int64
-        )
-        dgl_graph.edata[feat] = edge_assigns
-
-    for feat in node_attrs:
-        node_assigns = torch.tensor(
-            [val[-1] for val in graph.nodes(data=feat)], dtype=torch.int64
-        )
-        dgl_graph.ndata[feat] = node_assigns
-
-    return dgl_graph
-
-
-def update_graph_with_vocab(graph_fn, features, vocab):
-    """
-    Given a networkx attribute (function) and features update it with a vocab if possible.
-    If it cannot be updated, the features should already be numerical features.
-    Inputs:
-        - graph_fn: a networkx graph function (describing nodes or edges)
-        - features: the feature from the function that should be updated
-        - vocab: A dict mapping text to int
-    """
-
-    for feature_name in features:
-        curr_vocab = None
-        if feature_name in vocab:
-            curr_vocab = vocab[feature_name]
-        for graph_item in graph_fn(data=feature_name):
-            feature = graph_item[-1]
-            idx = graph_item[0]
-
-            if feature_name in vocab:
-                assert feature in curr_vocab
-                update_networkx_feature(
-                    graph_fn, idx, f"{feature_name}_idx", curr_vocab[feature]
-                )
-            else:
-                assert isinstance(
-                    feature, int
-                ), f"{(feature_name, feature)} is not an int"
-
-
-def update_networkx_feature(graph_fn, idx, feature_name, feature):
-    graph_fn[idx][feature_name] = feature
